All-platform make pass
