+++ proxy/tun/dev/dev.go
@@ -2,6 +2,7 @@ package dev
 
 import "github.com/google/netstack/tcpip/stack"
 
+// TunDevice is cross-platform tun interface
 type TunDevice interface {
 	Name() string
 	URL() string
